fixed related Grade packages and implemented Percent Grade logic

+++ src/main/java/com/nicat/storebonus/dtos/request/GradeRuleRequest.java
@@ -6,7 +6,7 @@ public record GradeRuleRequest(
         Long gradeId,
         Long positionId,
         Long marketId,
-        Double bonusPercent,
+        BigDecimal bonusPercent,
         BigDecimal amount
 ) {
 }
